Fixed some importing issues with the Python to GroMEt pipeline. Moved 'script_functions.py' into the util directory.
